// src/activity-log/types/action-type.types.ts
export enum ActionType {
  login = 'Logged into your account ',
  created_ad_account = 'Created new ad account ',
  deleted_ad_account = 'Deleted ad account ',
  update_ad_account = 'Updated ad account ',
  create_new_screen = 'Created new screen layout ',
  update_screen = 'Updated screen ',
  delete_screen = 'Deleted screen ',
  create_new_campaign = 'Created new campaign ',
  update_campaign = 'Updated campaign ',
  delete_campaign = 'Deleted campaign ',
  create_new_ad = 'Created new ad ',
  update_ad = 'Updated ad ',
  delete_ad = 'Deleted ad ',
  deleted_ad_company = 'Deleted ad company ',
  update_ad_company = 'Updated ad company ',
  created_ad_company = 'Created ad company ',
  update_widget_list = 'Updated Widget List ',
  update_widget_configuration = 'Updated Widget Configuration ',
  save_widget_configuration = 'Saved Widget Configuration ',
  create_widget_list = 'Created Widget List ',
  delete_widget_list = 'Deleted Widget List ',
  delete_widget_configuration = 'Deleted Widget Configuration ',
  activate_account = 'Activated account ',
  update_user = 'Updated  ',
  update_user_privilege = 'Updated user privilege of ',
  update_user_status = 'Update user status ',
  signUp = 'Signed up ',
  update_locale_data = 'Updated localedata ',
  update_username = 'Updated username ',
  deactivate_account = 'Deactivated account ',
  delete_account = 'Deleted account ',
  logout = 'Logged out',
  added_new_widget = 'Added new widget ',
  removed_widget = 'Removed widget ',
  add_user_to_organization = 'Added user to organization ',
}